refactor(voice): share GenAI message helpers
