+++ python/langsmith/integrations/livekit/_helpers.py
@@ -9,12 +9,15 @@ from __future__ import annotations
 
 import json
 import logging
-from collections.abc import Mapping
 from typing import Any, Optional
 
 from opentelemetry.util.types import AttributeValue
 
-from langsmith._internal.voice._helpers import try_parse_json_object
+from langsmith._internal.voice._helpers import (
+    build_assistant_tool_call_message,
+    build_tool_message,
+    try_parse_json_object,
+)
 from langsmith._internal.voice.translated_span import TranslatedSpan
 
 logger = logging.getLogger(__name__)
@@ -45,129 +48,6 @@ def get_content_attribute(tspan: TranslatedSpan, name: str) -> Optional[Attribut
     if pii_key in tspan.attributes:
         return tspan.attributes[pii_key]
     return tspan.attributes.get(f"lk.{name}")
-
-
-def _json_list(raw: object) -> Optional[list[object]]:
-    if isinstance(raw, str):
-        try:
-            raw = json.loads(raw)
-        except (ValueError, RecursionError):
-            return None
-    return raw if isinstance(raw, list) else None
-
-
-def _json_text(value: object) -> str:
-    return value if isinstance(value, str) else json.dumps(value)
-
-
-def _message_from_parts(role: str, parts: list[object]) -> list[dict[str, object]]:
-    content: list[dict[str, object]] = []
-    text_parts: list[str] = []
-    tool_calls: list[dict[str, object]] = []
-    tool_results: list[dict[str, object]] = []
-    for raw_part in parts:
-        if not isinstance(raw_part, Mapping):
-            continue
-        part: Mapping[object, object] = raw_part
-        kind = part.get("type")
-        part_content = part.get("content")
-        transcript = part.get("transcript")
-        uri = part.get("uri")
-        call_id = part.get("id")
-        tool_name = part.get("name")
-        if kind == "text" and isinstance(part_content, str):
-            text = part_content
-            content.append({"type": "text", "text": text})
-            text_parts.append(text)
-        elif (
-            kind == "blob"
-            and part.get("modality") == "audio"
-            and isinstance(transcript, str)
-        ):
-            content.append({"type": "text", "text": transcript})
-            text_parts.append(transcript)
-        elif (
-            kind == "uri"
-            and part.get("modality") == "image"
-            and isinstance(uri, str)
-            and uri
-        ):
-            content.append({"type": "image_url", "image_url": {"url": uri}})
-        elif (
-            kind == "tool_call"
-            and role == "assistant"
-            and isinstance(call_id, str)
-            and isinstance(tool_name, str)
-        ):
-            tool_calls.append(
-                {
-                    "id": call_id,
-                    "type": "function",
-                    "function": {
-                        "name": tool_name,
-                        "arguments": _json_text(part.get("arguments", {})),
-                    },
-                }
-            )
-        elif (
-            kind == "tool_call_response"
-            and role == "tool"
-            and isinstance(call_id, str)
-            and "response" in part
-        ):
-            tool_results.append(
-                build_tool_message(_json_text(part["response"]), tool_call_id=call_id)
-            )
-    messages: list[dict[str, object]] = []
-    if content or tool_calls:
-        message: dict[str, object] = {"role": role}
-        message["content"] = (
-            "\n".join(text_parts) if len(text_parts) == len(content) else content
-        )
-        if tool_calls:
-            message["tool_calls"] = tool_calls
-        messages.append(message)
-    return messages + tool_results
-
-
-def build_messages_from_gen_ai(raw: object) -> Optional[list[dict[str, object]]]:
-    """Convert LiveKit 1.8+ GenAI message parts to OpenAI-shaped messages.
-
-    ``None`` means missing/malformed input, allowing legacy event fallback;
-    an empty list is authoritative. Unknown parts are skipped, without fetching
-    media or logging conversation content.
-    """
-    items = _json_list(raw)
-    if items is None:
-        return None
-    messages: list[dict[str, object]] = []
-    for raw_item in items:
-        if not isinstance(raw_item, Mapping):
-            continue
-        item: Mapping[object, object] = raw_item
-        role = item.get("role")
-        parts = item.get("parts")
-        if not isinstance(role, str) or role not in (
-            "system",
-            "developer",
-            "user",
-            "assistant",
-            "tool",
-        ):
-            continue
-        if isinstance(parts, list):
-            messages.extend(_message_from_parts(role, parts))
-    return messages
-
-
-def build_system_messages_from_gen_ai(
-    raw: object,
-) -> Optional[list[dict[str, object]]]:
-    """Read the separate ``gen_ai.system_instructions`` text-part array."""
-    parts = _json_list(raw)
-    if parts is None:
-        return None
-    return _message_from_parts("system", parts)
 
 
 def normalize_provider(raw: Any) -> Optional[str]:
@@ -318,38 +198,6 @@ def flatten_lk_attributes_to_ls_metadata(
         ):
             flat[name] = list(v)
     return flat
-
-
-def build_tool_message(
-    content: str,
-    *,
-    tool_call_id: Optional[str] = None,
-    name: Optional[str] = None,
-) -> dict[str, object]:
-    """Build a ``tool`` result message, with its call id / name when present."""
-    msg: dict[str, object] = {"role": "tool", "content": content}
-    if tool_call_id:
-        msg["tool_call_id"] = str(tool_call_id)
-    if name:
-        msg["name"] = str(name)
-    return msg
-
-
-def build_assistant_tool_call_message(
-    call_id: str, name: str, arguments: str
-) -> dict[str, object]:
-    """Build an assistant message containing one tool call."""
-    return {
-        "role": "assistant",
-        "content": "",
-        "tool_calls": [
-            {
-                "id": call_id,
-                "type": "function",
-                "function": {"name": name, "arguments": arguments},
-            }
-        ],
-    }
 
 
 def build_message_from_event(role: str, event: Any) -> dict:
